:recycle: Separate dog cuddle animation from state
